First Interation

+++ controllers/cards.js
@@ -1,43 +1,46 @@
 const Card = require('../models/card');
+const ForbiddenError = require('../errors/ForbiddenError');
+const ValidationError = require('../errors/ValidationError');
+const NotFoundError = require('../errors/NotFoundError');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cardsData) => res.send(cardsData))
-    .catch(() => {
-      res.status(500).send({ message: 'На сервере произошла ошибка' });
-    });
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.send({ data: card }))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        return res.status(400).send({ message: err.message });
-      }
-      return res.status(500).send({ message: 'На сервере произошла ошибка' });
-    });
+    .catch(() => {
+      throw new ValidationError('Переданы неверные данные при создании карточки');
+    })
+    .catch(next);
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+module.exports.deleteCard = (req, res, next) => {
+  const userId = req.user._id;
+  Card.findById(req.params.cardId)
     .orFail(() => {
       const error = new Error('NotFound');
       throw error;
     })
-    .then((card) => res.send({ data: card }))
-    .catch((err) => {
-      if (err.message === 'NotFound') {
-        return res.status(404).send({ message: 'Запрашиваемая карточка не найдена' });
-      } if (err.name === 'CastError') {
-        return res.status(400).send({ message: 'Передан некорректный id карточки' });
+    .then((card) => {
+      if (card.owner.toString() !== userId) {
+        throw new ForbiddenError('Чужие карточки нельзя удалять');
       }
-      return res.status(500).send({ message: 'На сервере произошла ошибка' });
-    });
+      Card.findByIdAndRemove(req.params.cardId)
+        .then((newCard) => res.status(200).send({ data: newCard }))
+        .catch(next);
+    })
+    .catch((err) => {
+      throw new NotFoundError(err.message);
+    })
+    .catch(next);
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId,
     { $addToSet: { likes: req.user._id } },
     { new: true, runValidators: true })
@@ -48,15 +51,14 @@ module.exports.likeCard = (req, res) => {
     .then((likes) => res.send({ data: likes }))
     .catch((err) => {
       if (err.message === 'NotFound') {
-        return res.status(404).send({ message: 'Запрашиваемая карточка не найдена' });
-      } if (err.name === 'CastError') {
-        return res.status(400).send({ message: 'Передан некорректный id карточки' });
+        throw new NotFoundError(err.message);
       }
-      return res.status(500).send({ message: 'На сервере произошла ошибка' });
-    });
+      throw new ValidationError('Переданы неверные данные при лайке карточки');
+    })
+    .catch(next);
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId,
     { $pull: { likes: req.user._id } },
     { new: true, runValidators: true })
@@ -67,10 +69,9 @@ module.exports.dislikeCard = (req, res) => {
     .then((likes) => res.send({ data: likes }))
     .catch((err) => {
       if (err.message === 'NotFound') {
-        return res.status(404).send({ message: 'Запрашиваемая карточка не найдена' });
-      } if (err.name === 'CastError') {
-        return res.status(400).send({ message: 'Передан некорректный id карточки' });
+        throw new NotFoundError(err.message);
       }
-      return res.status(500).send({ message: 'На сервере произошла ошибка' });
-    });
+      throw new ValidationError('Переданы неверные данные при дизлайке карточки');
+    })
+    .catch(next);
 };
